docs: transition v2 docs from beta to stable state
- Drop beta banner, header widget, and index callout
- Rename V2 tab and default /docs redirect to v2
- Unpin install snippets, codemod, and static CDN from prerelease versions
