+++ apps/docs/app/layout.config.tsx
@@ -1,4 +1,3 @@
-import { V2BetaHeaderWidget } from '@/components/header-v2-widget'
 import { Logo } from '@/components/ui-blocks/logo'
 import { Widget3Icon } from '@solar-icons/react/dynamic/widget-3'
 import type { BaseLayoutProps } from 'fumadocs-ui/layouts/shared'
@@ -17,7 +16,7 @@ export const baseOptions: BaseLayoutProps = {
     links: [
         {
             text: 'Documentation',
-            url: '/docs/v1',
+            url: '/docs/v2',
             active: 'nested-url',
         },
         {
@@ -31,10 +30,6 @@ export const baseOptions: BaseLayoutProps = {
 
 export const homeOptions: BaseLayoutProps = {
     ...baseOptions,
-    nav: {
-        ...baseOptions.nav,
-        children: <V2BetaHeaderWidget />,
-    },
 }
 
 export const docsOptions: BaseLayoutProps = {
