+++ apps/docs/content/docs/v1/meta.json
@@ -1,6 +1,6 @@
 {
     "title": "V1",
-    "description": "Documentation for Solar Icons v1, the current stable release.",
+    "description": "Documentation for Solar Icons v1, the previous stable release.",
     "icon": "HistoryIcon",
     "root": true,
     "pages": ["index", "getting-started", "packages", "icon-reference", "advanced", "community"]
