+++ apps/docs/content/docs/v2/community/changelog.mdx
@@ -4,7 +4,7 @@ description: Release history for Solar Icons packages.
 icon: ClipboardListIcon
 ---
 
-## v2.0.0 (beta)
+## v2.0.0
 
 Unified major version across all packages. First release with a single package per framework, CSS variable cascade, built-in duotone, and stroke width control.
