+++ apps/docs/content/docs/v2/index.mdx
@@ -6,11 +6,6 @@ icon: HandShakeIcon
 
 import { Cards, Card } from 'fumadocs-ui/components/card'
 
-<Callout type="warn">
-    Solar Icons v2 is in beta. [Report issues on
-    GitHub](https://github.com/saoudi-h/solar-icons/issues).
-</Callout>
-
 Solar Icons gives you 1,246 distinct icons in six styles: **Bold**, **Linear**, **Outline**, **BoldDuotone**, **LineDuotone**, and **Broken**. With 7,476 total variations.
 
 ## Frameworks
