+++ apps/docs/content/docs/v2/migration-to-v2/react-perf.mdx
@@ -14,7 +14,7 @@ The [V2 codemod](/docs/v2/migration-to-v2/codemod) can rewrite the common React-
 
 ```bash
 npm uninstall @solar-icons/react-perf
-npm install @solar-icons/react@2.0.0-beta.0
+npm install @solar-icons/react
 ```
 
 ## Component naming
