+++ apps/docs/content/docs/v2/packages/angular.mdx
@@ -5,7 +5,7 @@ icon: angular
 ---
 
 ```package-install
-@solar-icons/angular@2.0.0-beta.0
+@solar-icons/angular
 ```
 
 Standalone components with signal-based inputs. `Solar`-prefixed PascalCase component names with attribute selectors on `<svg>`.
