+++ apps/docs/content/docs/v2/packages/js.mdx
@@ -5,7 +5,7 @@ icon: js
 ---
 
 ```package-install
-@solar-icons/js@2.0.0-beta.0
+@solar-icons/js
 ```
 
 ## Using a CDN (No Bundler)
@@ -27,7 +27,7 @@ You can use it directly in the browser using a modern CDN like [esm.sh](https://
                 createIcons,
                 HeartBoldIcon,
                 AccessibilityLineDuotoneIcon,
-            } from 'https://esm.sh/@solar-icons/js@2.0.0-beta.0'
+            } from 'https://esm.sh/@solar-icons/js'
 
             createIcons({
                 icons: {
