+++ apps/docs/content/docs/v2/packages/nuxt.mdx
@@ -5,7 +5,7 @@ icon: nuxt
 ---
 
 ```package-install
-@solar-icons/nuxt@2.0.0-beta.0
+@solar-icons/nuxt
 ```
 
 ## Installation
