+++ apps/docs/content/docs/v2/packages/react-native.mdx
@@ -5,7 +5,7 @@ icon: react
 ---
 
 ```package-install
-@solar-icons/react-native@2.0.0-beta.0 react-native-svg
+@solar-icons/react-native react-native-svg
 ```
 
 `react-native-svg` is a peer dependency. Install it manually if your project does not already use it.
