+++ apps/docs/content/docs/v2/packages/react.mdx
@@ -5,7 +5,7 @@ icon: react
 ---
 
 ```package-install
-@solar-icons/react@2.0.0-beta.0
+@solar-icons/react
 ```
 
 ## Import patterns
