+++ apps/docs/content/docs/v2/packages/solid.mdx
@@ -5,7 +5,7 @@ icon: solid
 ---
 
 ```package-install
-@solar-icons/solid@2.0.0-beta.0
+@solar-icons/solid
 ```
 
 ## Import patterns
