+++ apps/docs/content/docs/v2/packages/static.mdx
@@ -5,7 +5,7 @@ icon: static
 ---
 
 ```package-install
-@solar-icons/static@2.0.0-beta.0
+@solar-icons/static
 ```
 
 **What you can accomplish:**
@@ -114,16 +114,12 @@ To import an SVG as a string (e.g., for templating):
         <CodeBlockTabsTrigger value="Vite">Vite</CodeBlockTabsTrigger>
     </CodeBlockTabsList>
     <CodeBlockTab value="Webpack">
-```js
-// Specific icon style import
-import heartIcon from '@solar-icons/static/dist/icons/bold/heart.svg'
-```
+        ```js // Specific icon style import import heartIcon from
+        '@solar-icons/static/dist/icons/bold/heart.svg' ```
     </CodeBlockTab>
     <CodeBlockTab value="Vite">
-```js
-// Specific icon style import heartIcon from
-'@solar-icons/static/dist/icons/bold/heart.svg?raw'
-```
+        ```js // Specific icon style import heartIcon from
+        '@solar-icons/static/dist/icons/bold/heart.svg?raw' ```
     </CodeBlockTab>
 </CodeBlockTabs>
