+++ apps/docs/content/docs/v2/packages/svelte.mdx
@@ -5,7 +5,7 @@ icon: svelte
 ---
 
 ```package-install
-@solar-icons/svelte@2.0.0-beta.0
+@solar-icons/svelte
 ```
 
 Requires Svelte 5.0+. Uses runes (`$props`, `$derived`, `$bindable`).
