+++ apps/docs/content/docs/v2/packages/vue.mdx
@@ -5,7 +5,7 @@ icon: vue
 ---
 
 ```package-install
-@solar-icons/vue@2.0.0-beta.0
+@solar-icons/vue
 ```
 
 ## Import patterns
